requirements.txt
discord.py==2.3.2
yt-dlp>=2024.01.01
python-dotenv
PyNaCl
